User repo testing & other changes
Contains comprehensive testing of the user repo functions, including negative and
edge case testing.
Changes the operations of the user repo entirely into I-User. It is better to keep the
the models to mongo implementation and not leak their inner operations into the application
for example save() method.
Added the update() method to repository, which should replace the save().