Add a robust logging system
Logging is injected into all the classes and methods, supports different levels which are
logging to file, if the production logger is used.
Each file filters out only logs of that type, meaning that to see combined logs the files will need
to be combined. That is the choice of logging approach at the moment, but may need to
introduce a logging level which will aggregate all logs into one file.
Also added a dev logger which can be used in development, defaults to debug level and
logs everything into the console. No file logging.