Minor refactor & fix tests
Got rid of the abstract methods for repositories, everything should be preserved
in interfaces unless absolutely necessary.
Added a dummy logger to the testing environment, it was needed, after all.
No need to pollute the console and create files with error logs when things break.
Logging is tested on its own. Potentially we want to make that dummy logger into a
mock and test that logging works, but at the moment that adds a lot of complexity.
Changed some names, makes more sense to me.