add the view engine (ejs)
Add the support for views, shared layout, styles and other static files, etc.
The building of typescript into the dist/ directory and the whole microservice architecture
that I know will come to haunt me forced me to resort to a hacky way of copying all the views
and static files to dist directory and aggregating them under one dist/assets folder.
This will surely be a treat to manage if this application grows in complexity.
On another note - tests pass, but not always. Some asynchronous voodoo is going on.
So far from what I can tell it could be:
- server not shutting down properly in view engine tests
- server not having enough time to release the port before the next instance is brought up
- Need smarter file copyer, sometimes it's erroring out. Not sure why not all the time
- The test pre-setup needs to be more controlled. Chaos is happening at the moment.

Besides all that, I'm pretty happy with how this turned out.
Still hate building HTML.