[WIP] change logger to support module labels
Logging is certainly a neat thing, but it gets a bit overwhelming and confusing when the console is full
of messages from different places. It would be nice to show where a log is coming from.
It seems to be a bit more complex of a problem than I thought at first.
The problems:
- Creating a new instance of logger in current architecture creates new transports and files
- Passing a logger through a chain of calls forces to have one instance of a logger
- In my current attack at the problem, a logger instance can craete another logger instance, but this has its own downsides
  - If I forget to override the name of the logger downstream, the files will log with the name of their callers
  - Ugly
  - Writes to the same file and if the files are removed doesn't seem to recreate because of reused transports

Ideas:
- Logging factory in config provider?
- Singleton logging factory?
  This one I don't like much because then I can't dependency inject
  I guess I could configure that on startup...