Complete the Logging system
Well, this was a much longer journey than I expected, but here it is.
Logging, in its intended form will re-use transforms for creating different
Winston logger instances. Each class can create its own logger with its own label
which will be integrated into the output message.
I couldn't get it to print nicely, so for now we are getting it as a part of the JSON
dump at the end of the string.
Majority of the problems came from being able to write efficient tests, where I had
to add and remove transports on the fly, re-configure the Logger singleton, etc.
Definitely not necessary and definitely a HUGE WASTE OF MY TIME! But at the end I settled on a
compromise, thanks sanity. Definitely not something an insane person would say.